文章发布与标签

+++ blog/core/src/main/java/com/snow/blog/core/web/controller/ArticleController.java
@@ -3,10 +3,12 @@ package com.snow.blog.core.web.controller;
 import com.snow.blog.core.repository.entity.Article;
 import com.snow.blog.core.service.IArticleService;
 import com.snow.blog.core.util.validator.ArticleValidator;
+import com.snow.blog.core.web.controller.support.SearchArticleCondition;
+import com.snow.lib.result.ResultUtil;
 import com.snow.lib.result.ResultVO;
-import com.snow.lib.result.ResultVOUtil;
 import com.snow.security.core.repository.entity.User;
 import org.apache.commons.lang3.StringUtils;
+import org.hibernate.validator.constraints.NotEmpty;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -32,58 +34,55 @@ public class ArticleController {
 
     // 获取所有文章-分页
     @GetMapping("/page")
-    public ResultVO getArticlePage(
-            @NotNull String userId,
+    public ResultVO getArticlesPage(
+            @NotEmpty String userId,
             @PageableDefault(sort = {"createTime"}, direction = Sort.Direction.DESC) Pageable pageable,
             @AuthenticationPrincipal User user) {
-        if (StringUtils.isBlank(userId)) {
-            userId = user.getUserId();
-        }
         Page<Article> page = articlesService.getArticlePage(userId, user == null ? null : user.getUserId(), pageable);
-        return ResultVOUtil.page(page);
+        return ResultUtil.page(page);
     }
 
     // 搜索所有文章-分页
-    @GetMapping("/page/search")
-    public ResultVO getArticleSearchPage(
-            @NotNull String userId,
-            String content,
+    @PostMapping("/page/search")
+    public ResultVO searchArticles(
+            @RequestBody SearchArticleCondition condition,
             @PageableDefault(sort = {"createTime"}, direction = Sort.Direction.DESC) Pageable pageable,
             @AuthenticationPrincipal User user) {
-        if (StringUtils.isBlank(userId)) {
-            userId = user.getUserId();
-        }
-        Page<Article> page = articlesService.getArticleSearchPage(userId, user == null ? null : user.getUserId(), content, pageable);
-        return ResultVOUtil.page(page);
+        Page page = articlesService.searchArticleByCondition(condition, null == user ? null : user.getUserId(), pageable);
+        return ResultUtil.page(page);
     }
-
 
     // 获取指定文章
     @GetMapping("/{articleId}")
     public ResultVO getArticleByArticleId(@NotNull @PathVariable("articleId") String articleId, @AuthenticationPrincipal User user) {
         Article article = articlesService.getArticleById(articleId, user == null ? null : user.getUserId());
-        return ResultVOUtil.success(article);
+        return ResultUtil.success(article);
     }
 
     // 保存文章
     @PostMapping
-    public ResultVO saveArticle(@RequestBody Article article, @NotNull @AuthenticationPrincipal User user) {
-        Assert.isTrue(StringUtils.equals(user.getUserId(),article.getUserId()), "操作受限");
+    public ResultVO saveArticle(@RequestBody Article article, @AuthenticationPrincipal User user) {
+        if (null == user) {
+            return ResultUtil.unauthorized();
+        }
         ArticleValidator.validate(article);
         if (null == article.getPersonal()) {
             article.setPersonal(false);
         }
         Article result = articlesService.saveArticle(article, user == null ? null : user.getUserId());
-        return ResultVOUtil.success(result);
+        return ResultUtil.success(result);
     }
 
     // 删除文章
     @DeleteMapping
-    public ResultVO deleteArticle(@RequestBody Article article,@NotNull @AuthenticationPrincipal User user) {
-        Assert.isTrue(StringUtils.equals(user.getUserId(),article.getUserId()), "操作受限");
+    public ResultVO deleteArticle(@RequestBody Article article, @AuthenticationPrincipal User user) {
+        if (null == user) {
+            return ResultUtil.unauthorized();
+        }
+        Assert.isTrue(StringUtils.equals(user.getUserId(), article.getUserId()), "操作受限");
         ArticleValidator.articleExist(article);
         articlesService.deleteArticle(article);
-        return ResultVOUtil.success();
+        return ResultUtil.success();
     }
 
 }
